edx-55 created mixin to get contact information to reduce code replication (DRY)
